+++ sdk/master_board_sdk/example/example_pd.cpp
@@ -21,9 +21,9 @@ int main(int argc, char **argv)
 	int cpt = 0;
 	double dt = 0.001;
 	double t = 0;
-	float kp = 5.;
-	float kd = static_cast<float>(0.1);
-	float iq_sat = 4.0;
+	double kp = 5.;
+	double kd = 0.1;
+	double iq_sat = 4.0;
 	double freq = 0.5;
 	double amplitude = M_PI;
 	double init_pos[N_SLAVES * 2] = {0};
@@ -94,7 +94,7 @@ int main(int argc, char **argv)
 
 					// Use the current state as target for the PD controller.
 					robot_if.motors[i].SetCurrentReference(0.);
-					robot_if.motors[i].SetPositionReference(static_cast<float>(init_pos[i]));
+					robot_if.motors[i].SetPositionReference(init_pos[i]);
 					robot_if.motors[i].SetVelocityReference(0.);
 
 					t = 0;	//to start sin at 0
@@ -118,8 +118,8 @@ int main(int argc, char **argv)
 
 					if (robot_if.motors[i].IsEnabled())
 					{
-						float ref = static_cast<float>(init_pos[i] + amplitude * sin(2 * M_PI * freq * t));
-						float v_ref = static_cast<float>(2. * M_PI * freq * amplitude * cos(2 * M_PI * freq * t));
+						double ref = init_pos[i] + amplitude * sin(2 * M_PI * freq * t);
+						double v_ref = 2. * M_PI * freq * amplitude * cos(2 * M_PI * freq * t);
 
 						robot_if.motors[i].SetCurrentReference(0.);
 						robot_if.motors[i].SetPositionReference(ref);
